+++ classes/Dependency/Exception.php
@@ -1,3 +1,5 @@
 <?php defined('SYSPATH') or die('No direct script access.');
 
-class Dependency_Exception extends Kohana_Dependency_Exception {}
+class Dependency_Exception extends Kohana_Dependency_Exception {
+
+}