+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Improve the exception messages provided for easier debugging
 * Fix build and drop support for PHP <=5.5